refactor(runtime): extract model module from agent_loop
The model-identifier helper cluster moves into `agent_loop/model.rs`:

- `strip_provider_prefix` (pub) — strips `provider/` or `provider:`
  prefix from the catalog model id, then for providers that demand
  `org/model` form (OpenRouter, Together, Fireworks, Replicate,
  HuggingFace) backfills via `normalize_bare_model_id`.
- `apply_session_model_override_to_manifest` (pub) — per-session
  `model_override` apply path from #4898; called by
  `kernel::agent_execution::execute_llm_agent`.
- `needs_qualified_model_id` (pub(super)) — provider allowlist for the
  qualified-id rule above.
- `normalize_bare_model_id` (private) — bare-name → `org/model`
  lookup table.
- `DEFAULT_CONTEXT_WINDOW` (pub(super), `#[allow(dead_code)]`) — kept
  as the documented authoritative value referenced by
  `docs/architecture/message-history-trimming.md`.
- `UNKNOWN_MODEL_CONTEXT_WINDOW` (pub(super)) — 8K fallback used by
  the main loop's context-window resolution path (#3349).
- `stable_prefix_mode_enabled` (pub(super)) — manifest metadata check
  for the stable-prefix experiment flag.

The in-file `mod session_model_override_tests` (7 tests) travels with
`apply_session_model_override_to_manifest` so the test/function pair
stays co-located.

mod.rs re-exports the two `pub` items via
`pub use self::model::{...}` so kernel and channel callers keep
resolving the same `librefang_runtime::agent_loop::strip_provider_prefix`
path. `STABLE_PREFIX_MODE_METADATA_KEY` is removed from mod.rs's
import block (consumed only by `stable_prefix_mode_enabled`, which is
now in model.rs). The big `mod tests` block in mod.rs picks up
`needs_qualified_model_id` via the existing test-only
`use super::model::*` pattern.

Behavior-preserving. mod.rs: 13,349 -> 13,096 LOC (-253). cargo check
+ cargo clippy --lib both clean.

Refs #3710
